Dev 0.1.0

+++ atom-service/atom-service-operator/atom-service-operator-consumers/src/main/java/com/lamp/atom/service/operator/consumers/controller/ServiceController.java
@@ -31,8 +31,8 @@ public class ServiceController {
     PortCreatingFunction portCreatingFunction;
 
     @PostMapping("/getServicePort")
-    public Integer getServicePort(@RequestBody String ip) {
-        return portCreatingFunction.getPort(ip);
+    public Long getServicePort(@RequestBody String ip) {
+        return portCreatingFunction.getPort(ip).longValue();
     }
 
 }
